## What's New?
- Math for M, K augmented operators. We can now compute the modal matrices for the system.
- New element types: `JunctionNode` and `LumpedMass` as well as easier support for massless elements.

## Architectural changes
Separated the large `MBS` class into a `TopologyHandler` and `System` class for dealing with graph creation and traversal and modal computation respectively.

## Removed
- Extended state vectors. Reverted to the regular 12x1 state vectors as allowing the addition of external forces at this stage was overly complicated and didn't fit the current architecture
- Computation of static solution in `System` directly. A direct consequence of the removal of extended state vectors, this feature will instead be implemented when external forces are computed in the global dynamics of the ramp + rocket system.

+++ src/dynaramp/common/vecmath.py
@@ -19,7 +19,7 @@ def skew_sym_mat(vec: VectorLike) -> Matrix:
     :return:
     """
 
-    vec = np.array(vec)
+    vec = np.array(vec, dtype=np.float64)
     # Verify shape
     match vec.shape:
         # "Flat" vector
@@ -40,4 +40,4 @@ def skew_sym_mat(vec: VectorLike) -> Matrix:
             logger.error(err_msg)
             raise ValueError(err_msg)
 
-    return ssv
+    return ssv.astype(np.float64)
